refactor: update references to EN 1993-1-1:2022 in formulas 8.48, 8.49, 8.50, and 8.51
