--- blueprints/codes/eurocode/en_1993_1_1_2025/chapter_8_ultimate_limit_state/__init__.py
@@ -1,1 +1,0 @@
-"""Module containing all formulas from 1993-1-1+C2+A1:2025: Chapter 8 - Ultimate limit state."""
